Definition Module Migration
Going to move the Definition module out of the base NBTify project, as I don't think it necessarily fits the purpose as to what NBTify is meant for. NBTify is meant to be like a JSON module equivalent, but for NBT, and inherently SNBT. Making type definitions from/for NBT structures is an amazing feature, but it isn't really part of the NBT reading or writing process, so I don't think it's mainly related to it. I'm going to move it over to the Region-Types project instead, at least for the meantime.