Lightweight Generic Shadowing!
Got the format generic merging kind of working! Interestingly, this fixes the property types on the `NBTData` class itself, rather than adjusting what the values are, as they are passed into the constructor.

I was considering changing the implementations of `NBTData` type generics (for things like `read()`) to combine both the data generic, and the options generic, into a single one which is just a plain `NBTData` type, which you would define both of those two within that `NBTData` type specifically, rather than as two parameters on the function. However, I don't think I like this approach as much, since it makes you need to specify a full `NBTData` object every time, rather than only the two parameters individually.

I also think I'm going to hold off on the `build()` function implementation, as this mostly mitigates what that would be useful for/fix.

I tried doing this setup a few days ago I think, but it didn't work, because I tried doing it with `U[prop] extends unknown ? Prop : U[prop]`, or something like that. That doesn't work (as I thought) because it will match to `unknown` more than just for `unknown`, which I hadn't expected.

https://github.com/microsoft/TypeScript/issues/40855 (Not sure if this is related, but it helped with refining my searches)
https://stackoverflow.com/questions/68681526/why-cant-a-generic-member-function-in-a-class-implementing-an-interface-take-an (Noticing the use of `T extends {}` made me curious on trying it, as `{}` means any non-primitive type I think)

I think this article helps cover it a bit.
https://javascript.plainenglish.io/interviewer-please-tell-the-difference-between-object-and-object-in-typescript-5817b7810f61

I'm also re-thinking having the `NBTData` format options inheritance, as that seems to be not easy to type out, and it's not really too helpful to add on the type level either.

Can't remember if I mentioned it, but I want to work on the module exports structures too, so you can either import NBTify as a namespace (my preferred way of choice, for most instances), or to more easily import specific items under named exports. Rather than leaving this up to the user to add in their own aliases to make things easier to use, I want to add double-name exports for things, like `read()` to also be accessible though something like `readNBT()`, which would look nice when you don't want to use it as `NBT.read()`. You could get the choice as to which you like more :) Same thing with the other exports, `parseSNBT()` maybe, `stringifySNBT()` (I don't really like that, not sure about either of these), and `writeNBT()`. The other ones, like `NBTData` are already prefixed by nature, so I will leave those as-is.

Not sure if I'm still sold in keeping the reader/writer classes for each module inside each module, I'm deciding if it makes sense to move those into dedicated modules for themselves or not. I want to make it easier to see what each bit of the code is for, without having to go though a bunch of things to find what you're looking for, since the project is getting pretty big now.

Aah, I also want to get esbuild going eventually! Having a full-minimized build for the package bundle would be great to have out of the box. Maybe that doesn't make as much sense for npm though, I think I read that somewhere too. Maybe I should just leave that to the bundler when using this in the browser, since this works in ESM already too :)