Removing NBTData Shadowing
I think this isn't really too dangerous of a change, mostly because now that `NBTData` objects have format-based generics too, you can be certain that the 'type' of your `NBTData` object matches what it actually holds in it's property values. So, if you want an explicit format to be used in your new `NBTData` construct, you must be explicit about what that one should use.

I like this. I think I originally added the shadowing behavior to help counteract making the `NBTData` properties read-only, and I thought that would cause too much of a roadblock if you wanted to change only a set amount of properties. I think a better implementation to manage that might be to instead just use the `FormatOptions` object specifically when reading or writing to and from full NBT files. This effectively does the same thing, and it helps move some of the complexity away from the `NBTData` class, which should in theory probably be a fairly simple medium for representing the content it holds.