Return Await + Shadowing Bug (Was needed)
Ooop, I didn't test that code before pushing it, part of the reading code relied on this functionality. What a goof. Going back on the commit from yesterday.

Was thinking about it, and realized that I can remove the `result` variable in this case, and instead use `return await`. Here's an article about that which I remembered.

https://jakearchibald.com/2017/await-vs-return-vs-return-await/

Sucks about Google Domains :/