Compression Headers Revert
Bringing back the original compression header checks for the formats which can be checked. This makes things a bit simpler to run because you can figure out the compression format ahead of time, before needing to try decompressing with that given format.

Since there isn't a way to analytically discern between an uncompressed and `deflate-raw` compressed file directly, the simplest way to check which one it is, is to just try parsing as each, and see which one it is.

So with combining these two methods, I think this allows everything to work the same in terms of performance, as since the beginning, but now you also have `deflate-raw` auto-detection support too! This is the case because it will check for the header-ed formats first, and update the compression type based on that. If those aren't found, then it will try reading as uncompressed first (the most common, compared to `deflate-raw`), then it will try `deflate-raw` if the uncompressed doesn't work. If none of the formats work, then the auto-detect error will return whatever the error would be for the uncompressed file read.

To expand on that, the default format that it tries to read/shows error messages for (when no formats are specified), will be for trying to read a named, big endian, uncompressed, non-Bedrock Level file.