Root ListTag: Initial Support
This is the initial implementation and script I wrote to generate a test file that I can demo things out with. Next I'm going to save a local copy of this to the repo, then I can run the full test runner script on it to demo out the implementation on top of that.

This is how I've been doing the other formats up until now too, the only difference is that I can automate the part of ensuring that the currently-supported formats stay working great :)

Listening to 'The Edge of Forever', was listening to 'Wine and Pickles' again, 'Inhale' is a cool song for sure.

I still have to set up SNBT support, I'm going to do that alongside the test runner, after committing these initial changes. Once again, commit what you have working to start with, before you goof it up! haha

#23