Root ListTag: Fully Symmetrical!
Saved the local copy of the TAG_List root tagged NBT data, and added support for the missing SNBT write handling for TAG_List root tags. With the new file part of the test runner, and in conjunction with the new SNBT write handling, it now passes the test too! Super cool. And with how the SNBT parsing module is written, it happened to be already ready to support TAG_List root tags! Thinking about how that main parser function works though, I think I do need to narrow the types that it allows for the main SNBT string, as I want it to only allow what a full SNBT string would allow from the root, now any SNBT string passed into it. Hence, I think right now it would allow you to pass in just a simple TAG_Byte string in, and it would work correctly. This isn't what the full binary NBT spec allows though, so I don't want SNBT to accept it either.

Oh yeah, and this new format option/data structure reminded me that I should add this to the `ridiculous.nbt` file, along with `deflate-raw` compression. That will make it even more of a monstrous file to test other implementations' support for (NBTify too).

Yeah, not like that kind of file would ever really be out in the wild, but making sure that all of those formats work together can help strengthen that you know your logic for those are all working as expected, that your library can discern between all of the different crazy things you throw at it. That's my goal with this one, I want NBTify to handle whatever you try to do with it!

#23

https://minecraft.fandom.com/wiki/NBT_format#:~:text=In%20Bedrock%20Edition%2C%20the%20root%20tag%20in%20an%20NBT%20file%20can%20sometimes%20be%20a%20%C2%A0TAG_List%20instead%20of%20a%20%C2%A0TAG_Compound.
https://wiki.bedrock.dev/nbt/nbt-in-depth.html#:~:text=You%20also%20need%20to%20pay%20attention%20to%20the%20root%20element%20in%20the%20file%2C%20i.e.%20the%20list%20or%20compoud