Ridiculous Upgrade
Updated the new testing logic to now organize the file listing irrespective of each file's capitalization; I like that more :)

This search helped out a lot with this, looks like there's a standard way to do this now, neat!

https://stackoverflow.com/questions/8996963/how-to-perform-case-insensitive-sorting-array-of-string-in-javascript/40390844#40390844

Upgraded the `ridiculous.nbt` file to now use `deflate-raw` for the compression format, a ListTag for the root tag, and I made the Bedrock Level header negative, to show the signed-nature of the value. I think it's supposed to be signed? I'd like to think that it should be unsigned, since it's a version number, but it doesn't really name that specifically anywhere.

There might be a secret easter egg entry in there too, hehe. It might be about a certain Terrible Lizard.