Compression Piping Revert, Revert...
Wow, this was an interesting one! Turns out my current compression code doesn't work in the browser, discovered this while trying to update to this latest version in Dovetail.

https://github.com/Offroaders123/Dovetail/issues/1

Realized that I made the Bedrock Level header version for `ridiculous.nbt` a signed int, oops.

I really want to find out whether the header is signed or not, my implementation currently goes through signed *and* unsigned for values, so mine isn't very structurally sound, come to think of it.