Compression Uncaught Error Handling
Wow, this was an interesting one! I think it may have to do with how the backing implementation of the Compression Streams API is implemented in the browser(s) vs Node. Turns out after all, the error is still partially around the `writer.write()` and `writer.close()` calls.

So, in Node, awaiting both calls works, because they have a `.then()`-able return value, while in Chrome, Safari, and Firefox, they don't ever return when chained with either `.then()`, or used with `await`, hence why it would hang with what would seem like an errorless result. In Node however though, these do eventually resolve, with just a `void` return value.

And interestingly enough, not sure if it has to do with the compression backing implementation, or the handling of Promise rejections and thread sleeping (I don't remember what that's called, but having to do with when the task ends, hiding any errors which haven't rejected by the time the task is completed) (I think it is that second one, I'll explain), but those errors don't occur in Node, but they do in the browser.

I think this is because the console closes before the Promises have rejected, meaning they simply get hidden from failing in the function calls.

I think this can be related to something similar to that of Jake Archibald's article about `for await`, and how it can't properly catch unhandled Promises, as they aren't caught by the current thread pool (not sure if that's the proper term either)

https://jakearchibald.com/2023/unhandled-rejections/

While searching possible fixes to help figure this out, seeing the extra `.catch()` handlers made it click a little more for me, and it help further discover where these ghost errors were bubbling up from!

https://stackoverflow.com/questions/37624322/uncaught-in-promise-undefined-error-when-using-with-location-in-facebook-gra

Overall, I think the main issue is that `writer.write()` and `writer.close()` don't properly have `.then()`-able resolutions in browser(s), while they do in Node. If I could use `await` on these calls, then I could ensure that these errors properly get bubbled back up to the main function, which could properly send them as feedback up to the caller, which can decide how to handle them as they need.

In the meantime, I think I simply have to rely on the errors that would fire after these, because these can't be successfully caught by the main containing function. So I think using a ghost `.catch()` method on them will be the only fix for it, unfortunately.

Maybe I'm missing something about how you should use `write()` and `close()` and maybe they shouldn't be awaited. I'd think that would be key to catching errors properly in situations like these though, no?

https://github.com/Offroaders123/Dovetail/issues/1

For a while, I thought this had to do with my other changes to the Read module (Root ListTags, `return await` try-catch calls, `deflate-raw` auto-detect handling), but turns out it was just the error handling within the Compression module itself, for when piping through a compression stream would have an error, and it simply wan't being cast to the parent scope correctly, because it wasn't being called with `await` (Because it doesn't work in the browser).

Ooh, and the part I said I'd go back to! I think this was happening in Node actually, but only when the thread was running long enough to allow the errors to bubble up to become Uncaught. When I implemented the auto-detect without header checks change, that was using try-catch for each of the compression format types, and errors would magically cause the NEXT compression format to error out, and I couldn't figure out why. Turns out it *was* because Node was exiting in every other case, before those un-awaited Promises were erroring out (because the *header* based checks were faster than when the header-less checks were running), meaning the header-less checks ran long enough for the uncaught errors in the previous auto-detect compression test would bubble up to the main scope as uncaught, and I associated this as only happening with the `deflate-raw` format, because it would only happen with that one, because the other checks would run fast enough to finalize before the errors would bubble up in time.