Declarative Declarations
Related to the last commit, I'm deciding to use interfaces and objects directly to represent the NBT, rather than using interfaces and classes. I don't want inheritance to get in the way of the structure representations. It also makes it much easier for working with existing NBT files/structures, because you don't have to build a class from the declarative object (the data), instead the object is what you're using to work with the data.

So, I guess I'm taking the same route as how NBTify's primitives themselves work, I hopefully want to have the smallest level of abstractions on top of the data, only representing things using primitives, and working with those primitives using functional implementations.