ListTag Explicit Generic
Realized now that types are represented more strictly for NBT structures (like with ListTagUnsafe and such), maybe it could make more sense to always need the user to specify the type for all of the items in the list tag. I think this can help with making things more narrow type-wise, rather than just falling back to `Tag` as a default. I think making the type's user need to decide this distinction is a key thing to making NBT representation more accurate, as they may not use the generic if they don't know that they 'should' supply one.