Implicity Strict!
Found out that functions can't have required generics, unlike classes and types.

But I did just find out that you can use `never` in place or `unknown` in cases like these.

It's a little less explanatory than `unknown` I think, which is unfortunate, but it does enable the correct type behavior intended from using `T extends RootTag = unknown`.

This in return successfully makes it so it's up to the user to decide to type the data on their own, or traverse into it unsafely using `any`. So, this is still fully backwards compatible with code before these changes, except that if you were relying on `any` previously, now you just have to specify that in use of the `NBTData` type itself, being that the first generic parameter is now required.

#32

https://stackoverflow.com/questions/51173191/typescript-require-generic-parameter-to-be-provided/63413428#63413428
https://stackoverflow.com/questions/75183307/args-never-vs-args-unknown-in-typescript/75185028#75185028