Primitive Wrapper Objects: Take Two
After talking some more with @HoldYourWaffle, he thinks using primitive wrapper objects/classes instead of plain primitives can provide for more UX in terms of instance methods and things like that, rather than a functional approach that manipulates the primitives themselves.

I'm not fully committed to going back to using these classes instead of the primitives setup, since I think it's easier to work with the data as raw JS values than with full wrapper objects, but I think it would be important to try out this setup again now that I have had some more help and guidance on setting things up properly (not wonky haha).

In terms of typing things out using the wrapper classes as types (the standard lib thing, like with 'Region-Types'), `builder-class.ts` in the test folder is the closest thing as to what it would look like to make interface shapes for existing NBT, using the classes as the types instead.

This commit also tries out the `Tag | undefined` key definition for the `CompoundTag` type! This semi-fixes the issue with `RootTag` being the default return type from `NBT.read()` and the like, in that it will still show hallucinated properties on `RootTag`, but you won't be able to access them directly, because they are both from an index signature, and half of their type are from their `ListTag` declaration, and the other half as `Tag | undefined`.

So while things like `Array.concat` show up on `RootTag`, you can't fully read from them, nor call them as methods, because at the type level they are in union with the type `Tag | undefined`. So `Array.length` would show up as `number | Tag | undefined`.

HoldYourWaffle thinks this being more explicit is more helpful to the user, but I think seeing `any` makes more sense just because it doesn't make you think those properties are there. Most NBT data already isn't scoped inside a root `ListTag`, so seeing `Array.concat` as the first thing in your autocomplete doesn't seem helpful, but more misleading to what it *is* returning.

I think it's great to try out new things though, so exploring new ways to work on it I'm all ears for!

I think having methods on the NBT tag types would be helpful to the user, yes, but it does make the rest of representing NBT shapes and structures much heavier to handle, which I don't think is a fair trade-off. At least, I'm trying to write things more functionally-driven now, so having functionality based on instance methods seems to go against that.