Optional CompoundTag Keys
Brought over from the test 'Primitive Wrapper Objects: Take Two' branch, now `undefined` is an acceptable entry on `CompoundTag` key/value pairs!

This change also makes the default `RootTag` return type for various function calls more usable, but I'm also going to roll back that test in this commit. Having `RootTag` as the default return type makes sense in terms of what it is at the type level, but it isn't as user-friendly when doing things with it specifically. So the default return type will once again be `any`, which feels more synonymous with how you would work with `JSON.parse()` or `Response.prototype.json()`, which NBTify is semi-trying to mirror in some aspects. #32

I think I need to rework the unsafe/safe `ListTag` and `CompoundTag` handling again, as now with having `undefined` being an ok value at the type level, there needs to be a better line of distinction of where it is present in the `CompoundTag` type when it is passing through the library. It gets removed when passed through `fromCompoundUnsafe()`, but the return type `CompoundTag` still lists it as a possible value for keys of `CompoundTag`. So for `undefined` at least, I need to at minimum move the handling for that into places where `CompoundTag`s are traversed directly, and put `undefined` handling there, since it is allowed on the type now, unlike before.

*Stringify and Write modules use '!' to counteract this inconsistency at the moment*

Oh yeah, and `NBTData` still requires a first parameter generic, I'm not going to revert that one, as it does help a lot with having the user be more explicit about their types when possible. They have to choose whether or not to provide their own type, or just use `any` if they don't know their types yet. If functions could require generics too, then I wouldn't default the return types to `any`, as requiring a generic would be the best use case for the UX of the return types. But I understand why that doesn't make as much sense, so it's ok :)

Got lots of music the last few days! They have all been OUTSTANDING!!

Haken - Affinity 🕹️
Devin Townsend - Lightwork 🐙
Cleft - Whale Bone 🐋
Mike Keneally - Inkling (More from the Scambot 2 Sessions) 🦂🦂
Devin Townsend - Silent Night ❄️