Unsafe Tag Handling (pt. 2)
Wrote a test to make sure that invalid properties will be skipped when traversed for the Write and Stringify modules, and discovered I missed some handling in the write module, which didn't match against unsupported keys inside of `CompoundTag` objects.

Realized that using `NBTData` as a class actually still uses `RootTag` as the default content value, so I'm going to reset that one back to `any` as well. It was nice to have it required for use as a type, but it looks like you also can't require a generic when using a class in the runtime-sense (not as a type). So, I think most all of the changes from
the last few commits have been reverted. Sometimes the nicer options of multiple evils doesn't work out, and that's ok. I like the UX for this setup, it feels ergonomically-expandable! (No idea if that means anything really, but feels fancy) #32

I think NBTify is very solid as of now, you can do a lot of things with it! Going to see what other things I can do to break it, in hopes of making it stronger 💪