Improved Tipes + To Organize
Was surprised that 'tipes' isn't a real word, thought it would be taken for something already. Any time I think of a weird name of a mispelling on purpose for the memes, I always Google it to make sure it's not something I don't mean it to be lol. Did that ever happen to your when you were little? You use a word because you thought it had a funny reaction from people, or that it just sounded cool, but then it's bad and you're not supposed to use it? That happened a few times for me before, and I've been on edge from doing it again accidentally 😂

Listening to Psychedelic Porn Crumpets, such a great album! (High Visceral {Part 1})