SNBT Unit Tests
Going to add some unit tests for explicitly parsing SNBT from the source, to ensure that string escaping and things like that are working as expected.

While working on this, realized I forgot to re-add indentation during the recompilation step, and that made me think while writing this description; should the source test file have indentation, or be minified? Probably both would be best, now that I think about it.

Gonna have to make `ridiculous.nbt` more ridiculous, now that I also have characters to escape in SNBT-land.

Noticed that having the `RootTag`-like return types for the function APIs set to `RootTag` explicitly rather than any does appear to help distinguish things correctly apart, rather than `any`. Now I'm starting to see what @HoldYourWaffle was proving with having that as the default return type. I may actually move the types back from `any` again, mostly because of the type elision when both are `any`. So the `RootTag` type itself might not be useful for the plain type it represents, but it definitely is worth having for preventing type elision when working with other unknown, or rather, `any` values.

#26
#32