Strict Generics + SNBT Character Escaping
Changes from this and the last few commits:

Brought back the strict handling for the API generics, now they will indeed default to `RootTag` afterall. I'm finally convinced! haha. I discussed this re-change in the previous commit. #32

Updated the Node typings to the latest version. I don't think it likely changed anything for NBTify, did it more just for maintenance.

`CompoundTag` keys now accept optional values! This is both to allow more types which have to be described at the lib level (some keys are optional in the Entity interface), and to prevent accidental use of the values when accessing them from the direct `RootTag` type, if you don't explcitly provide a shape to type check against yourself.

Unsafe tag handling is now once again part of the writing and stringifying steps individually, rather than being dedicated types and used through explicit functions. The use of checking this in both seemed to be redundant to do in both places, but I think it's more composable-like to remove this as an abstraction, since they didn't both follow the same use requirements, so you had to iterate over the data once to verify it, then again to use it. I like it better where you have to iterate over the data one single time, doing the manual tag type validation there instead.

SNBT now properly accounts for the necessary escaped characters required by the spec. Going to look into improving the use of optional quotes for `StringTag` values while in SNBT. I'm debating whether the writer should also emit non-quoted values where possible. The drawback to that is needing to check whether the values isn't a crossover with what would accidentally be parsed as another SNBT tag type. So `"1b"` wouldn't validate as `1b`, because `1b` would mean a `ByteTag` value, meaning the type would be translated incorrectly. That's the kind of things I want to watch out for. Because of the intricacies that come with that, I'm curious if the Wiki is mis-quoted there, but maybe not. It could be manageable, so I'll look more into whether it seems like a viable/realistic option to support.

Added a standalone SNBT test to the unit test runner!