No-Dual-Build
Going back on making a dual-build setup for both ESM and CommonJS. Was already deciding on this before signing off for the night during last night's coding session, but it was getting a bit too late to add the reverted changes for this.

Before making this commit I wanted to do some more research to see what other takes on this topic would be. It seems now that Node has had proper support for ESM for some time now, it's more logical to go with pure-ESM for newer packages, and I think NBTify could count under that category.

https://www.youtube.com/watch?v=TXcFnsY5aqQ
https://www.youtube.com/watch?v=lMWUqWKEGgQ
https://www.reddit.com/r/node/comments/121a1wa/esm_vs_dual_package/
https://levelup.gitconnected.com/transpiling-typescript-into-double-packages-commonjs-esm-d6b62cfc851c
https://preciselab.io/simplest-tutorial-for-esm-commonjs-package-creators/