Internal Inspect Types
While looking at the docs, for updating some things to work with the latest TypeScript version (gonna come in the next commit), found out what I was looking for a while back; making things hidden from the build types! This fixes the issue of needing to mark the `CustomInspect` getters on the primitive wrapper objects as `private`, as you can just hide them from the build output using JSDoc `@internal` statements instead. This also opens the door to the other thing, enabling unused code checking in the TSConfig! This wasn't an option before while needing to mark the methods as private, because none of the code for the library itself uses those methods, only Node does. They had to be marked private/not included in the build output types because I didn't want `@types/node` to be a dependency for the project, just to have typings for a feature that the user won't use directly.

https://www.typescriptlang.org/tsconfig#stripInternal