Escaped Quotes
Discovered a bug in the new text escaping logic for SNBT which doesn't handle escaped quotes in the parsing stage.

I set up a test to validate that this is working both ways. Currently it isn't there at passing the test, but it appears to be working for the demo, to start with?

'The Mistakes' (Mike Keneally) is such a killer album, can't wait to get it and 'The Future Bites' (Steven Wilson), both of which were the listens for tonight so far.