Parse/Stringify Text Fix
Differences in indentation were breaking the test check, so I had to ensure that there was one more level to where it would properly format it to match the control data. So, it's not ideal because the 'control' data for the test isn't the raw data, but the only other option is for the SNBT file to already be formatted the same as how it will be outputted. But I don't necessarily want that though, because I want to make sure the indentation parsing works with these escapes too. Maybe they should be separate tests; that's probably the correct way to cover those cases accordingly.