Output Format Flags
Working on the setup to allow the CLI to pipe to stdout, so you can use the NBT content directly with other programs using piping/teeing!

This works as of this release now!
```sh
nbtify ./ridiculous.snbt --name= --endian=big --compression=deflate
```

I think it's something I'll have to do like this to use stdout?
https://stackoverflow.com/questions/4976466/difference-between-process-stdout-write-and-console-log-in-node-js