Error Visibility
Realized that NBTify's internal argument type check handling works great for this! I was debating removing that in favor of just TS types instead, leaving the runtime to be honestly a bit of a wild west. That's not reliable for cases just like this though. It does bloat up the internal code a bit, which was partly why I was thinking of removing it, but it is definitely worthwhile now that I've seen it really helps with things like this. I don't have to add explicit type checks in the CLI, because the API calls in NBTify themselves type check the values for you, so it's something that can be strongly relied on :)

The only other thing I might try is to make NBTify's union types into enums maybe, and it can be something that the library can more strongly and consistently check against, since right now I'm using literal strings a lot of the time, and `instanceof` checks. I want to make those a bit less extensive and splintered across the codebase.

(This was gonna be part of the next commit, but I'm gonna come back to this later; https://stackoverflow.com/questions/20585729/how-to-detect-if-nodes-process-stdout-is-being-piped)