RootTagLike
Trying another middle-ground solution to allow values that don't specifically extends the `RootTag` and friends' types. Essentially, the issue is that you can't use `as RootTag` on a plain object, and I don't want to force consumers to add `extends CompoundTag` or things like that to their shapes, because that will add unintended index signature key values to show up when accessing values off the type, which I discovered while working on Region-Types.

#19
#28
#32