Run-Time List Item Type Check (SNBT)
Added checks to the SNBT parser and stringifier that ensures ListTag items are all of the same type. Just like the reader and writer classes, the type is determined by the first item in the array.

Did things a bit different to implement this one, and it ended up working out really well! This time I started things off with updating the test to fail if the SNBT stringifier worked successfully when it shouldn't, then updated the code to make the test pass (or rather, to properly throw when it encounters an improperly-typed list item). Then I did the same thing for the parsing stage, made the test fail if it parsed successfully, then added the proper throw calls when the parsed items weren't all of the same type as the first item in the array. This makes me feel stronger about the checking here, because I passed it data that *shouldn't* work, and it only passes the test if the program doesn't accept that malformed test demo. Great way to do things like that! Now I just want to refactor the test to make it much less ugly, lol. But I also don't want to break it's behavior in doing that, it could cause issues if it lets unchecked changes pass though to the library if they aren't being validated correctly, compared to this version.

Man, I really still have to rework the SNBT parse module, it's so hacky at the moment. At least, it feels that way. It works great though, so I'd rather have that over anything else!

#14