Stronger FormatOptions
Demoing this out, trying to use the general 'Format'-like structure throughout the codebase, rather than having distinct options interfaces for each module.

This isn't fully complete for the Read module though, I just realized I think it's because the current `ReadOptions` isn't attached to the `NBTDataOptions` generic, hence why passing in a specific format to `read()` doesn't cast that type to the result `NBTData` as a type generic, which I do want it to do.

#29 (Having to do with generic 'merging' kind of, at least maybe rather calling it 'generic propagation' in this case)