VarInt Reset
Got some more intel from PremiereHell on how Bedrock's VarInt works, and I wanted to start fresh with a new implementation again.

I also decided to open the VarInt demo file in HexEd.it, and compare it to a decompressed version of the regular `bigtest.nbt` (Edit: I also saved this edit as little endian, which makes it easier to tell the difference between the two files' bytes more accurately). I noticed that it uses VarInt a lot more than I originally thought, and it's for everything Int-like and Long-like, not only those tag types themselves. So things like the length of a StringTag are encoded with VarInt, same with the length of CompoundTag entries' key names, all of that kind of thing. So I'm going to diff where things aren't the same in that regard, then implement it based on that. I think it's more something tangible to try and understand with, versus just a spec that I have to equate to how I think it should work.

This is that spec, and it's definitely helpful, I think I just didn't know where it should be implemented/used more specifically.
https://wiki.vg/VarInt_And_VarLong

Cool tip, `open` (and probably `xdg-open`) work with PWAs! Dovetail is the default file handler for NBT files on my system, so I could do `open ./test/nbt/bigtest.nbt` and it opened right with Dovetail! Extra cool.