SNBT Float Types Formatting
Discovered a bug with my SNBT output! Realized that it's actually 'mandatory' (I think?) for Float and Double tags to have a decimal point. I thought it was a bug in Tnze's SNBT extension for VSCode, but it was my misunderstanding as to what was required for those tags when displayed/viewed/edited in SNBT.

I also just realized I need to add support for the `true` and `false` boolean keyword types. Those don't appear to be working either.

https://github.com/Tnze/snbt-vscode
https://static.wikia.nocookie.net/minecraft_de_gamepedia/images/e/e2/SNBT_data-Befehl.png/revision/latest
https://minecraft.wiki/w/NBT_format#SNBT_format

#36